fix: fetch topics on MFE load
