+++ src/discussions/learners/LearnerPostsView.jsx
@@ -1,4 +1,4 @@
-import React, {
+import {
   useCallback, useContext, useEffect, useMemo,
 } from 'react';
